ai_service/requirements.txt
torch
torchvision
timm
albumentations
opencv-python
numpy
pandas
scikit-learn
tqdm
faiss-cpu
fastapi
uvicorn
python-multipart
pillow
